+++ resolvers/index.js
@@ -2,7 +2,7 @@ const { Constructor, Constructors } = require("../data/Constructors/resolvers");
 const { Driver, Drivers } = require("../data/Drivers/resolvers");
 const { SeasonsResolver } = require("./Seasons");
 const { StatusResolver } = require("../data/Status/resolvers");
-const { ResultsResolver } = require("./Results");
+const { Results } = require("../data/Results/resolvers");
 const { QualiResolver } = require("./Qualifying");
 const { DriverStandings } = require("./DriverStandings");
 const { TeamStandingsResolver } = require("./TeamStandings");
@@ -13,7 +13,7 @@ const resolvers = {
     Constructors,
     Driver,
     Drivers,
-    Results: ResultsResolver,
+    Results,
     Seasons: SeasonsResolver,
     Status: StatusResolver,
     QualifyingResults: QualiResolver,
